[ELEMENTS][WL] move element typings into elements folder
