+++ src/components/elements/alert/types.ts
@@ -1,7 +1,8 @@
 import { AlertProps, AlertType } from "@lifesg/react-design-system";
-import { IFrontendEngineElementJsonSchema, TComponentOmitProps } from "../../frontend-engine";
+import { TComponentOmitProps } from "../../frontend-engine";
+import { IBaseElementSchema } from "../types";
 
-export interface IAlertSchema extends IFrontendEngineElementJsonSchema<"alert">, TComponentOmitProps<AlertProps> {
+export interface IAlertSchema extends IBaseElementSchema<"alert">, TComponentOmitProps<AlertProps> {
 	type: AlertType;
 	children: React.ReactNode;
 }
