+++ src/components/elements/index.ts
@@ -1,2 +1,3 @@
 export * from "./alert";
 export * from "./text";
+export * from "./types";
