+++ src/components/elements/wrapper/wrapper.tsx
@@ -7,8 +7,10 @@ import { TestHelper } from "../../../utils";
 import { useFormSchema } from "../../../utils/hooks";
 import * as FrontendEngineCustomComponents from "../../custom";
 import { ECustomElementType, ECustomFieldType } from "../../custom";
+import { EElementType } from "../../elements";
 import * as FrontendEngineFields from "../../fields";
-import { EElementType, EFieldType, TFrontendEngineFieldSchema } from "../../frontend-engine/types";
+import { EFieldType } from "../../fields";
+import { TFrontendEngineFieldSchema } from "../../frontend-engine/types";
 import { ERROR_MESSAGES } from "../../shared";
 import { ConditionalRenderer } from "./conditional-renderer";
 import { FieldWrapper } from "./field-wrapper";
